+++ upnp/src/ssdp/ssdp_device.c
@@ -302,7 +302,7 @@ static int NewRequestHandler(
 			*(RqPacket + Index));
 		rc = sendto(ReplySock,
 			*(RqPacket + Index),
-			strlen(*(RqPacket + Index)),
+			(int)strlen(*(RqPacket + Index)),
 			0,
 			DestAddr,
 			socklen);
